# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KATALA - Katalog & Reservasi Wisata Lampung API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket detail with QR code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/validate:
    post:
      operationId: validateTicket
      tags: [staff]
      summary: Validate a ticket by QR code (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketValidateInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketValidationResult"
        "404":
          description: Ticket not found
        "400":
          description: Already used or expired

  # ─── PAYMENTS ────────────────────────────────────────────
...
components:
  schemas:
...
    Destination:
      type: object
      required: [id, name, description, location, latitude, longitude, price, dailyQuota, category, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        price:
          type: number
        dailyQuota:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: ["number", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
...
    Reservation:
      type: object
      required: [id, userId, destinationId, visitDate, quantity, totalPrice, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        destinationId:
          type: integer
        visitDate:
          type: string
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled, expired]
        destination:
          $ref: "#/components/schemas/Destination"
        createdAt:
          type: string
...
    Ticket:
      type: object
      required: [id, reservationId, ticketCode, qrCodeData, status]
      properties:
        id:
          type: integer
        reservationId:
          type: integer
        ticketCode:
          type: string
        qrCodeData:
          type: string
        status:
          type: string
          enum: [valid, used, expired]
        usedAt:
          type: ["string", "null"]
        reservation:
          $ref: "#/components/schemas/Reservation"
        createdAt:
          type: string
    TicketValidateInput:
      type: object
      required: [ticketCode]
      properties:
        ticketCode:
          type: string
    TicketValidationResult:
      type: object
      required: [valid, message, ticket]
      properties:
        valid:
          type: boolean
        message:
          type: string
        ticket:
          $ref: "#/components/schemas/Ticket"

    # ─── PAYMENTS ──────────────────────────────────────────
...