# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KATALA - Katalog & Reservasi Wisata Lampung API
 * OpenAPI spec version: 0.1.0
 */
import type { Reservation } from "./reservation";
import type { TicketStatus } from "./ticketStatus";

export interface Ticket {
  id: number;
  reservationId: number;
  ticketCode: string;
  qrCodeData: string;
  status: TicketStatus;
  /** @nullable */
  usedAt?: string | null;
  reservation?: Reservation;
  createdAt?: string;
}